LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/sibling-layout-dependency-expected.txt

XX

FAIL Sibling style mutation assert_equals: expected "20" but got ""
FAIL Sibling style mutation, parent is affected assert_equals: expected "20" but got ""
FAIL Sibling style mutation, ancestor is affected assert_equals: expected "20" but got ""
FAIL Sibling text mutation assert_equals: expected "20" but got ""
